requirements.txt
blessed==1.20.0
botorch==0.8.5
contourpy==1.1.1
cycler==0.12.1
deap==1.4.1
enlighten==1.12.4
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
future==1.0.0
gpytorch==1.12
importlib-resources==6.4.0
jaxtyping==0.2.19
jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
linear-operator==0.4.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
multipledispatch==1.0.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
# opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==24.1
pandas==2.0.3
pillow==10.4.0
prefixed==0.7.1
pyparsing==3.1.2
pyro-api==0.1.2
pyro-ppl==1.9.1
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sympy==1.13.1
threadpoolctl==3.5.0
# torch==1.8.0+cu111
# torchaudio==0.8.0
# torchvision==0.9.0+cu111
tqdm==4.66.4
triton==3.0.0
typeguard==2.13.3
typing-extensions==4.12.2
tzdata==2024.1
wcwidth==0.2.13
# zellij==1.0.1
zipp==3.19.2
